+++ frontend/api/household.ts
@@ -1,6 +1,12 @@
 import { gql } from 'graphql-request'
 import { getGraphqlClient } from './client';
 
+const CREATE_HOUSE_MUTATION = gql`
+    mutation CreateHousehold($householdName: String) {
+        createHousehold(householdName: $householdName)
+    }
+`
+
 const ADDUSER_MUTATION = gql`
     mutation AssignHousehold($user: string, $household: string) {
         assignHousehold(userid: $user, houseid: $household)
@@ -12,6 +18,11 @@ const CHECKHOUSECODE_QUERY = gql`
         checkHouseCode(joinCode: $code)
     }
 `;
+
+export const createGroup = async (name: any) =>{
+    const client = await getGraphqlClient();
+    return client.request(CREATE_HOUSE_MUTATION, {name})
+}
 
 export const addUser = async (userID: any, houseID: any) => {
     const client = await getGraphqlClient();
